refactor(copyright): decompose tree_walk copyright module
Split the ~2.7k-line src/copyright/detector/tree_walk/copyright.rs into a
copyright/ directory module organized by tree-walk concern:

- prefix.rs: copyright/portions/not-copyrighted prefix recovery
- node_classify.rs: pure node-classification predicates and lookahead helpers
- absorb.rs: should_start_absorbing + collect_trailing_orphan_tokens and
  following-clause/year-range absorption
- merge.rs: cross-node copyright<->author and year-only<->copyrighted-by merges
- tree_nodes.rs: the primary extract_from_tree_nodes driver
- spans.rs: extract_from_spans / extract_copyrights_from_spans flat-leaf passes
- bare.rs: extract_bare_copyrights / extract_holder_is_name

Behavior-preserving: all moved logic is byte-for-byte identical aside from
import paths and visibility. Public paths are preserved via re-exports in
copyright/mod.rs (then tree_walk's `pub use copyright::*`); the test-only
helpers should_start_absorbing/collect_trailing_orphan_tokens stay reachable
through a crate-visible `absorb` module instead of an unused re-export. No
`.unwrap()` exists in the module; the single infallible `.expect("same_line
checked")` is preserved.

Verified with cargo build/clippy(-D warnings)/fmt, the copyright::detector lib
tests, and the copyright_golden suite; no golden expectations changed.
